+++ sdk/grpc/GrpcClient.cpp
@@ -108,7 +108,7 @@ GrpcClient::CreateIndex(const ::milvus::grpc::IndexParam& index_param) {
 }
 
 Status
-GrpcClient::Insert(::milvus::grpc::VectorIds& vector_ids, const ::milvus::grpc::InsertParam& insert_param) {
+GrpcClient::Insert(const ::milvus::grpc::InsertParam& insert_param, ::milvus::grpc::VectorIds& vector_ids) {
     ClientContext context;
     ::grpc::Status grpc_status = stub_->Insert(&context, insert_param, &vector_ids);
 
@@ -136,6 +136,23 @@ GrpcClient::GetVectorByID(const grpc::VectorIdentity& vector_identity, ::milvus:
     if (vector_data.status().error_code() != grpc::SUCCESS) {
         std::cerr << vector_data.status().reason() << std::endl;
         return Status(StatusCode::ServerFailed, vector_data.status().reason());
+    }
+
+    return Status::OK();
+}
+
+Status
+GrpcClient::GetIDsInSegment(const grpc::GetVectorIDsParam& param, grpc::VectorIds& vector_ids) {
+    ClientContext context;
+    ::grpc::Status grpc_status = stub_->GetVectorIDs(&context, param, &vector_ids);
+
+    if (!grpc_status.ok()) {
+        std::cerr << "GetIDsInSegment rpc failed!" << std::endl;
+        return Status(StatusCode::RPCFailed, grpc_status.error_message());
+    }
+    if (vector_ids.status().error_code() != grpc::SUCCESS) {
+        std::cerr << vector_ids.status().reason() << std::endl;
+        return Status(StatusCode::ServerFailed, vector_ids.status().reason());
     }
 
     return Status::OK();