+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Please mark all change in change log and use the issue from GitHub
 - \#853 - Support HNSW
 - \#910 - Change Milvus c++ standard to c++17
 - \#1204 - Add api to get table data information
+- \#1302 - Get all record IDs in a segment by given a segment id
 
 ## Improvement
 - \#738 - Use Openblas / lapack from apt install